install and upgrade ca-certificates.
centos 6 needs to have it's ca-certificates package upgraded so that it
can pull the epel/metalink from the epel repo that gets added from the
epel-release package. the epel-release package uses https in the
/etc/yum.repos.d/epel.repo. without updating the ca-certificates
packages we cannot negotiate an trusted ssl connection to the epel
remote repositories and therefore are not able to install the nginx
package.